Switch to oconfigure
Remove the existing "compat" machinery in favour of oconfigure.  This
simplifies the Makefile and allows for the removal of the compat files
and reallocarray shim.

While here, use WWWDIR instead of PREFIX for installwww (like other
bsd.lv projects) and add a distclean target.

The churn is in renaming compat.h -> config.h.